updated the imu logic, added actuators connection test, and refactored the master test script
